+++ src/tests/FlashOWare.Tool.Cli.Tests/Sdk/DotNet.cs
@@ -1,0 +1,11 @@
+namespace FlashOWare.Tool.Cli.Tests.Sdk;
+
+public sealed partial class DotNet
+{
+    private readonly DirectoryInfo _directory;
+
+    public DotNet(DirectoryInfo directory)
+    {
+        _directory = directory;
+    }
+}