+++ src/tests/FlashOWare.Tool.Cli.Tests/Sdk/DotNetNewTemplate.cs
@@ -1,0 +1,7 @@
+namespace FlashOWare.Tool.Cli.Tests.Sdk;
+
+internal enum DotNetNewTemplate
+{
+    Unspecified = 0,
+    AspNetCoreWebApiNativeAot = 1,
+}